+++ component/application_info.go
@@ -14,9 +14,9 @@
 
 package component
 
-// ApplicationStartInfo is the information that is logged at the application start and
+// BuildInfo is the information that is logged at the application start and
 // passed into each component. This information can be overridden in custom builds.
-type ApplicationStartInfo struct {
+type BuildInfo struct {
 	// Executable file name, e.g. "otelcol".
 	ExeName string
 
@@ -24,9 +24,9 @@ type ApplicationStartInfo struct {
 	Version string
 }
 
-// DefaultApplicationStartInfo returns the default ApplicationStartInfo.
-func DefaultApplicationStartInfo() ApplicationStartInfo {
-	return ApplicationStartInfo{
+// DefaultBuildInfo returns the default BuildInfo.
+func DefaultBuildInfo() BuildInfo {
+	return BuildInfo{
 		ExeName: "otelcol",
 		Version: "latest",
 	}