+++ component/receiver.go
@@ -59,8 +59,8 @@ type ReceiverCreateParams struct {
 	// component to be used later as well.
 	Logger *zap.Logger
 
-	// ApplicationStartInfo can be used by components for informational purposes
-	ApplicationStartInfo ApplicationStartInfo
+	// BuildInfo can be used by components for informational purposes
+	BuildInfo BuildInfo
 }
 
 // ReceiverFactory can create TracesReceiver and MetricsReceiver. This is the