+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Change `With*Unmarshallers` signatures in Kafka exporter/receiver (#2973)
 - Rename `marshall` to `marshal` in all the occurrences (#2977)
 - Remove `componenterror.ErrAlreadyStarted` and `componenterror.ErrAlreadyStopped`, components should not protect against this, Service will start/stop once.
+- Rename `ApplicationStartInfo` to `BuildInfo`
 
 ## 💡 Enhancements 💡